requirements.txt
Django==2.2.13
djangorestframework==3.12.1
django-widget-tweaks==1.4.8
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1